+++ examples/lt_ex_hw_wallet.c
@@ -20,7 +20,7 @@
 #define PING_MSG_SIZE 26
 
 /** @brief Attestation key for ECC slot 0. */
-uint8_t attestation_key[32]
+uint8_t attestation_key[TR01_CURVE_PRIVKEY_LEN]
     = {0x22, 0x57, 0xa8, 0x2f, 0x85, 0x8f, 0x13, 0x32, 0xfa, 0x0f, 0xf6, 0x0c, 0x76, 0x29, 0x42, 0x70,
        0xa9, 0x58, 0x9d, 0xfd, 0x47, 0xa5, 0x23, 0x78, 0x18, 0x4d, 0x2d, 0x38, 0xf0, 0xa7, 0xc4, 0x01};
 
@@ -470,7 +470,7 @@ static int session1(lt_handle_t *h)
     }
     LT_LOG_INFO("\tOK");
 
-    uint8_t dummy_key[32] = {0};
+    uint8_t dummy_key[TR01_SHIPUB_LEN] = {0};
     LT_LOG_INFO("Writing all pairing key slots (should fail):");
     for (uint8_t i = TR01_PAIRING_KEY_SLOT_INDEX_0; i <= TR01_PAIRING_KEY_SLOT_INDEX_3; i++) {
         LT_LOG_INFO("\tWriting pairing key slot %" PRIu8 "...", i);
@@ -536,7 +536,7 @@ static int session2(lt_handle_t *h)
     LT_LOG_INFO("Message received from TROPIC01:");
     LT_LOG_INFO("\t\"%s\"", recv_buf);
 
-    uint8_t dummy_key[32] = {0};
+    uint8_t dummy_key[TR01_CURVE_PRIVKEY_LEN] = {0};
     LT_LOG_INFO("Trying to store key into ECC slot %d (should fail)", (int)TR01_ECC_SLOT_0);
     ret = lt_ecc_key_store(h, TR01_ECC_SLOT_0, TR01_CURVE_ED25519, dummy_key);
     if (LT_L3_UNAUTHORIZED != ret) {
@@ -558,7 +558,7 @@ static int session2(lt_handle_t *h)
 
     uint32_t mcounter_value = 0x000000ff;
     LT_LOG_INFO("Initializing mcounter 0 (should fail)...");
-    ret = lt_mcounter_init(h, 0, mcounter_value);
+    ret = lt_mcounter_init(h, TR01_MCOUNTER_INDEX_0, mcounter_value);
     if (LT_L3_UNAUTHORIZED != ret) {
         LT_LOG_ERROR("Return value is not LT_L3_UNAUTHORIZED, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -566,7 +566,7 @@ static int session2(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Updating mcounter 0 (should fail)...");
-    ret = lt_mcounter_update(h, 0);
+    ret = lt_mcounter_update(h, TR01_MCOUNTER_INDEX_0);
     if (LT_L3_UNAUTHORIZED != ret) {
         LT_LOG_ERROR("Return value is not LT_L3_UNAUTHORIZED, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -574,7 +574,7 @@ static int session2(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Getting mcounter 0 (should fail)...");
-    ret = lt_mcounter_get(h, 0, &mcounter_value);
+    ret = lt_mcounter_get(h, TR01_MCOUNTER_INDEX_0, &mcounter_value);
     if (LT_L3_UNAUTHORIZED != ret) {
         LT_LOG_ERROR("Return value is not LT_L3_UNAUTHORIZED, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -637,8 +637,8 @@ static int session3(lt_handle_t *h)
 
     LT_LOG_INFO("Signing with attestation key which was updated through pairing key slot 1");
     uint8_t msg[] = {'a', 'h', 'o', 'j'};
-    uint8_t rs[64];
-    ret = lt_ecc_eddsa_sign(h, TR01_ECC_SLOT_0, msg, 4, rs);
+    uint8_t rs[TR01_ECDSA_EDDSA_SIGNATURE_LENGTH];
+    ret = lt_ecc_eddsa_sign(h, TR01_ECC_SLOT_0, msg, sizeof(msg), rs);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to sign, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -646,10 +646,10 @@ static int session3(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Reading ECC key slot %d...", (int)TR01_ECC_SLOT_0);
-    uint8_t slot_0_pubkey[64];
+    uint8_t ed25519_pubkey[TR01_CURVE_ED25519_PUBKEY_LEN];
     lt_ecc_curve_type_t curve;
     lt_ecc_key_origin_t origin;
-    ret = lt_ecc_key_read(h, TR01_ECC_SLOT_0, slot_0_pubkey, &curve, &origin);
+    ret = lt_ecc_key_read(h, TR01_ECC_SLOT_0, ed25519_pubkey, &curve, &origin);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to read ECC slot, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -657,7 +657,7 @@ static int session3(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Verifying with lt_ecc_eddsa_sig_verify()...");
-    ret = lt_ecc_eddsa_sig_verify(msg, 4, slot_0_pubkey, rs);
+    ret = lt_ecc_eddsa_sig_verify(msg, sizeof(msg), ed25519_pubkey, rs);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to verify, ret%s", lt_ret_verbose(ret));
         return -1;
@@ -699,7 +699,7 @@ static int session3(lt_handle_t *h)
 
     uint32_t mcounter_value = 0x000000ff;
     LT_LOG_INFO("Initializing mcounter 0...");
-    ret = lt_mcounter_init(h, 0, mcounter_value);
+    ret = lt_mcounter_init(h, TR01_MCOUNTER_INDEX_0, mcounter_value);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to initialize mcounter, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -707,7 +707,7 @@ static int session3(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Updating mcounter 0...");
-    ret = lt_mcounter_update(h, 0);
+    ret = lt_mcounter_update(h, TR01_MCOUNTER_INDEX_0);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to update mcounter, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -715,14 +715,14 @@ static int session3(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Getting mcounter 0...");
-    ret = lt_mcounter_get(h, 0, &mcounter_value);
+    ret = lt_mcounter_get(h, TR01_MCOUNTER_INDEX_0, &mcounter_value);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to get mcounter, ret=%s", lt_ret_verbose(ret));
         return -1;
     }
     LT_LOG_INFO("\tOK");
 
-    uint8_t dummy_key[32];
+    uint8_t dummy_key[TR01_CURVE_PRIVKEY_LEN];
     LT_LOG_INFO("Trying to store key into ECC slot %d (should fail)", (int)TR01_ECC_SLOT_0);
     ret = lt_ecc_key_store(h, TR01_ECC_SLOT_0, TR01_CURVE_ED25519, dummy_key);
     if (LT_L3_UNAUTHORIZED != ret) {