+++ examples/lt_ex_macandd.c
@@ -49,8 +49,8 @@
  */
 struct lt_macandd_nvm_t {
     uint8_t i;
-    uint8_t ci[MACANDD_ROUNDS * 32];
-    uint8_t t[32];
+    uint8_t ci[MACANDD_ROUNDS * TR01_MAC_AND_DESTROY_DATA_SIZE];
+    uint8_t t[LT_HMAC_SHA256_HASH_LEN];
 } __attribute__((__packed__));
 
 /**
@@ -85,18 +85,18 @@ static lt_ret_t lt_PIN_set(lt_handle_t *h, const uint8_t *PIN, const uint8_t PIN
 
     // Clear variable for released secret so there is known data (zeroes) in case this function ended sooner then secret
     // was prepared
-    memset(secret, 0, 32);
+    memset(secret, 0, TR01_MAC_AND_DESTROY_DATA_SIZE);
 
     // Variable used during a process of getting a encryption key k_i
-    uint8_t v[32] = {0};
+    uint8_t v[LT_HMAC_SHA256_HASH_LEN] = {0};
     // Variable used during a process of getting a encryption key k_i
-    uint8_t w[32] = {0};
+    uint8_t w[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
     // Encryption key
-    uint8_t k_i[32] = {0};
+    uint8_t k_i[LT_HMAC_SHA256_HASH_LEN] = {0};
     // Random secret
-    uint8_t s[32] = {0};
+    uint8_t s[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
     // Variable used to initialize slot(s)
-    uint8_t u[32] = {0};
+    uint8_t u[LT_HMAC_SHA256_HASH_LEN] = {0};
 
     // This organizes data which will be stored into nvm
     struct lt_macandd_nvm_t nvm = {0};
@@ -128,17 +128,18 @@ static lt_ret_t lt_PIN_set(lt_handle_t *h, const uint8_t *PIN, const uint8_t PIN
     nvm.i = MACANDD_ROUNDS;
     // Compute tag t = KDF(s, "0"), save into nvm struct
     // Tag will be later used during lt_PIN_check() to verify validity of secret
-    lt_hmac_sha256(s, 32, (uint8_t *)"0", 1, nvm.t);
+    lt_hmac_sha256(s, sizeof(s), (uint8_t *)"0", 1, nvm.t);
 
     // Compute u = KDF(s, "1")
     // This value will be sent through M&D sequence to initialize a slot
-    lt_hmac_sha256(s, 32, (uint8_t *)"1", 1, u);
+    lt_hmac_sha256(s, sizeof(s), (uint8_t *)"1", 1, u);
 
     // Compute v = KDF(0, PIN||A) where 0 is all zeroes key
-    lt_hmac_sha256((uint8_t*)"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00", 32, kdf_input_buff, PIN_size+add_size, v);
+    const uint8_t zeros[32] = {0};
+    lt_hmac_sha256(zeros, sizeof(zeros), kdf_input_buff, PIN_size + add_size, v);
 
     for (int i = 0; i < nvm.i; i++) {
-        uint8_t garbage[32] = {0};
+        uint8_t garbage[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
 
         // This call of a M&D sequence results in initialization of one slot
         LT_LOG_INFO("Doing M&D sequence to initialize a slot...");
@@ -170,15 +171,15 @@ static lt_ret_t lt_PIN_set(lt_handle_t *h, const uint8_t *PIN, const uint8_t PIN
 
         // Derive k_i = KDF(w, PIN||A)
         // This key will be used to encrypt secret s
-        lt_hmac_sha256(w, 32, kdf_input_buff, PIN_size + add_size, k_i);
+        lt_hmac_sha256(w, sizeof(w), kdf_input_buff, PIN_size + add_size, k_i);
 
         // Encrypt s using k_i as a key
         //
         // Warning: All s[n] are encrypted here using k_i[n] as a key, results are stored into ci[n].
         // Because size of the 'message s' is the same as a size of the 'key k_i', simple XOR is used here - discuss
         // more apropriate method with experts on cryptography.
-        for (int j = 0; j < 32; j++) {
-            *(nvm.ci + (i * 32 + j)) = k_i[j] ^ s[j];
+        for (unsigned int j = 0; j < TR01_MAC_AND_DESTROY_DATA_SIZE; j++) {
+            *(nvm.ci + (i * TR01_MAC_AND_DESTROY_DATA_SIZE + j)) = k_i[j] ^ s[j];
         }
     }
 
@@ -191,15 +192,15 @@ static lt_ret_t lt_PIN_set(lt_handle_t *h, const uint8_t *PIN, const uint8_t PIN
     }
 
     // Final secret is released to the caller
-    lt_hmac_sha256(s, 32, (uint8_t *)"2", 1, secret);
+    lt_hmac_sha256(s, sizeof(s), (uint8_t *)"2", 1, secret);
 
 // Cleanup all sensitive data from memory
 exit:
     memset(kdf_input_buff, 0, PIN_size + add_size);
-    memset(u, 0, 32);
-    memset(v, 0, 32);
-    memset(w, 0, 32);
-    memset(k_i, 0, 32);
+    memset(u, 0, sizeof(u));
+    memset(v, 0, sizeof(v));
+    memset(w, 0, sizeof(w));
+    memset(k_i, 0, sizeof(k_i));
 
     return ret;
 }
@@ -236,20 +237,20 @@ static lt_ret_t lt_PIN_check(lt_handle_t *h, const uint8_t *PIN, const uint8_t P
 
     // Clear variable for released secret so there is known data (zeroes) in case this function ended sooner then secret
     // was prepared
-    memset(secret, 0, 32);
+    memset(secret, 0, TR01_MAC_AND_DESTROY_DATA_SIZE);
 
     // Variable used during a process of getting a decryption key k_i
-    uint8_t v_[32] = {0};
+    uint8_t v_[LT_HMAC_SHA256_HASH_LEN] = {0};
     // Variable used during a process of getting a decryption key k_i
-    uint8_t w_[32] = {0};
+    uint8_t w_[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
     // Decryption key
-    uint8_t k_i[32] = {0};
+    uint8_t k_i[LT_HMAC_SHA256_HASH_LEN] = {0};
     // Secret
-    uint8_t s_[32] = {0};
+    uint8_t s_[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
     // Tag
-    uint8_t t_[32] = {0};
+    uint8_t t_[LT_HMAC_SHA256_HASH_LEN] = {0};
     // Value used to initialize Mac And Destroy's slot after a correct PIN try
-    uint8_t u[32] = {0};
+    uint8_t u[LT_HMAC_SHA256_HASH_LEN] = {0};
 
     // This organizes data which will be read from nvm
     struct lt_macandd_nvm_t nvm = {0};
@@ -297,7 +298,8 @@ static lt_ret_t lt_PIN_check(lt_handle_t *h, const uint8_t *PIN, const uint8_t P
     LT_LOG_INFO("\tOK");
 
     // Compute v’ = KDF(0, PIN’||A).
-    lt_hmac_sha256((uint8_t*)"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00", 32, kdf_input_buff, PIN_size + add_size, v_);
+    const uint8_t zeros[32] = {0};
+    lt_hmac_sha256(zeros, sizeof(zeros), kdf_input_buff, PIN_size + add_size, v_);
 
     // Execute w’ = MACANDD(i, v’)
     LT_LOG_INFO("Doing M&D sequence...");
@@ -309,7 +311,7 @@ static lt_ret_t lt_PIN_check(lt_handle_t *h, const uint8_t *PIN, const uint8_t P
     LT_LOG_INFO("\tOK");
 
     // Compute k’_i = KDF(w’, PIN’||A)
-    lt_hmac_sha256(w_, 32, kdf_input_buff, PIN_size + add_size, k_i);
+    lt_hmac_sha256(w_, sizeof(w_), kdf_input_buff, PIN_size + add_size, k_i);
 
     // Read the ciphertext c_i and tag t from NVM, decrypt c_i with k’_i as the key and obtain s_
 
@@ -317,25 +319,25 @@ static lt_ret_t lt_PIN_check(lt_handle_t *h, const uint8_t *PIN, const uint8_t P
     //
     // For getting s[n] again, they are decrypted here by XORing ci[n] with k_i[n]
     // Discuss more propriate method with experts on cryptography.
-    for (int j = 0; j < 32; j++) {
-        s_[j] = *(nvm.ci + (nvm.i * 32 + j)) ^ k_i[j];
+    for (unsigned int j = 0; j < sizeof(s_); j++) {
+        s_[j] = *(nvm.ci + (nvm.i * TR01_MAC_AND_DESTROY_DATA_SIZE + j)) ^ k_i[j];
     }
 
     // Compute tag t = KDF(s, "0x00")
-    lt_hmac_sha256(s_, 32, (uint8_t *)"0", 1, t_);
+    lt_hmac_sha256(s_, sizeof(s_), (uint8_t *)"0", 1, t_);
 
     // If t’ != t: FAIL
-    if (memcmp(nvm.t, t_, 32) != 0) {
+    if (memcmp(nvm.t, t_, sizeof(t_)) != 0) {
         ret = LT_FAIL;
         goto exit;
     }
 
     // Pin is correct, now initialize macandd slots again:
     // Compute u = KDF(s’, "0x01")
-    lt_hmac_sha256(s_, 32, (uint8_t *)"1", 1, u);
+    lt_hmac_sha256(s_, sizeof(s_), (uint8_t *)"1", 1, u);
 
     for (int x = nvm.i; x < MACANDD_ROUNDS - 1; x++) {
-        uint8_t garbage[32] = {0};
+        uint8_t garbage[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
 
         LT_LOG_INFO("Doing M&D sequence...");
         ret = lt_mac_and_destroy(h, x, u, garbage);
@@ -365,14 +367,14 @@ static lt_ret_t lt_PIN_check(lt_handle_t *h, const uint8_t *PIN, const uint8_t P
     LT_LOG_INFO("\tOK");
 
     // Calculate secret and store it into passed array
-    lt_hmac_sha256(s_, 32, (uint8_t *)"2", 1, secret);
+    lt_hmac_sha256(s_, sizeof(s_), (uint8_t *)"2", 1, secret);
 
 // Cleanup all sensitive data from memory
 exit:
     memset(kdf_input_buff, 0, PIN_size + add_size);
-    memset(w_, 0, 32);
-    memset(k_i, 0, 32);
-    memset(v_, 0, 32);
+    memset(w_, 0, sizeof(w_));
+    memset(k_i, 0, sizeof(k_i));
+    memset(v_, 0, sizeof(v_));
 
     return ret;
 }
@@ -403,7 +405,7 @@ int lt_ex_macandd(lt_handle_t *h)
     }
 
     // This variable stores secret which is released to the user after successful PIN check or PIN set
-    uint8_t secret[32] = {0};
+    uint8_t secret[TR01_MAC_AND_DESTROY_DATA_SIZE] = {0};
 
     // Additional data passed by user besides PIN - this is optional, but recommended
     uint8_t additional_data[]
@@ -419,7 +421,7 @@ int lt_ex_macandd(lt_handle_t *h)
 
     // Set the PIN and log out the secret
     LT_LOG("Setting the user PIN...");
-    ret = lt_PIN_set(h, pin, 4, additional_data, additional_data_size, secret);
+    ret = lt_PIN_set(h, pin, sizeof(pin), additional_data, additional_data_size, secret);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Failed to set the user PIN, ret=%s", lt_ret_verbose(ret));
         lt_session_abort(h);
@@ -428,7 +430,7 @@ int lt_ex_macandd(lt_handle_t *h)
     }
     LT_LOG_INFO("\tOK");
     char print_buff[PRINT_BUFF_SIZE];
-    ret = lt_print_bytes(secret, 32, print_buff, PRINT_BUFF_SIZE);
+    ret = lt_print_bytes(secret, sizeof(secret), print_buff, PRINT_BUFF_SIZE);
     if (LT_OK != ret) {
         LT_LOG_ERROR("lt_print_bytes failed, ret=%s", lt_ret_verbose(ret));
         return -1;
@@ -439,14 +441,14 @@ int lt_ex_macandd(lt_handle_t *h)
     LT_LOG_INFO("Doing %d PIN check attempts with wrong PIN...", MACANDD_ROUNDS);
     for (int i = 1; i < MACANDD_ROUNDS; i++) {
         LT_LOG_INFO("\tInputting wrong PIN -> slot #%d destroyed", i);
-        ret = lt_PIN_check(h, pin_wrong, 4, additional_data, additional_data_size, secret);
+        ret = lt_PIN_check(h, pin_wrong, sizeof(pin_wrong), additional_data, additional_data_size, secret);
         if (LT_FAIL != ret) {
             LT_LOG_ERROR("Return value is not LT_FAIL, ret=%s", lt_ret_verbose(ret));
             lt_session_abort(h);
             lt_deinit(h);
             return -1;
         }
-        ret = lt_print_bytes(secret, 32, print_buff, PRINT_BUFF_SIZE);
+        ret = lt_print_bytes(secret, sizeof(secret), print_buff, PRINT_BUFF_SIZE);
         if (LT_OK != ret) {
             LT_LOG_ERROR("lt_print_bytes failed, ret=%s", lt_ret_verbose(ret));
             return -1;
@@ -456,14 +458,14 @@ int lt_ex_macandd(lt_handle_t *h)
     LT_LOG_INFO("\tOK");
 
     LT_LOG_INFO("Doing Final PIN attempt with correct PIN, slots are reinitialized again...");
-    ret = lt_PIN_check(h, pin, 4, additional_data, additional_data_size, secret);
+    ret = lt_PIN_check(h, pin, sizeof(pin), additional_data, additional_data_size, secret);
     if (LT_OK != ret) {
         LT_LOG_ERROR("Attempt with correct PIN failed, ret=%s", lt_ret_verbose(ret));
         lt_session_abort(h);
         lt_deinit(h);
         return -1;
     }
-    ret = lt_print_bytes(secret, 32, print_buff, PRINT_BUFF_SIZE);
+    ret = lt_print_bytes(secret, sizeof(secret), print_buff, PRINT_BUFF_SIZE);
     if (LT_OK != ret) {
         LT_LOG_ERROR("lt_print_bytes failed, ret=%s", lt_ret_verbose(ret));
         return -1;