+++ src/lt_crc16.c
@@ -8,6 +8,8 @@
 
 #include "lt_crc16.h"
 
+#include "libtropic_common.h"
+
 /* Generator polynomial value used */
 #define LT_CRC16_POLYNOMIAL 0x8005
 
@@ -25,9 +27,10 @@ static uint16_t crc16_byte(uint8_t data, uint16_t crc)
 
     current_byte = data;
     crc ^= current_byte << 8;
-    i = 8;
+    i = 8;  // Iterate over every bit in a byte.
     do {
-        if (crc & 0x8000) {
+        if (crc
+            & 0x8000) {  // Highest bit set -> carry -> add generator polynomial of finite field used in CRC calulation.
             crc <<= 1;
             crc ^= LT_CRC16_POLYNOMIAL;
         }
@@ -55,7 +58,8 @@ uint16_t crc16(const uint8_t *data, int16_t len)
 void add_crc(void *req)
 {
     uint8_t *p = (uint8_t *)req;
-    uint16_t len = p[1] + 2;
+    // CRC is calculated from REQ_DATA, REQ_ID and REQ_LEN.
+    uint16_t len = p[TR01_L2_REQ_RSP_LEN_OFFSET] + TR01_L2_REQ_ID_SIZE + TR01_L2_REQ_RSP_LEN_SIZE;
 
     uint16_t crc = crc16(p, len);