+++ src/lt_hkdf.c
@@ -32,15 +32,15 @@ void lt_hkdf(uint8_t *ck, uint32_t ck_size, uint8_t *input, uint32_t input_size,
 {
     LT_UNUSED(nouts);
 
-    uint8_t tmp[32] = {0};
+    uint8_t tmp[LT_HMAC_SHA256_HASH_LEN] = {0};
     uint8_t one = 0x01;
 
     lt_hmac_sha256(ck, ck_size, input, input_size, tmp);
-    lt_hmac_sha256(tmp, 32, &one, 1, output_1);
+    lt_hmac_sha256(tmp, sizeof(tmp), &one, 1, output_1);
 
     uint8_t helper[33] = {0};
-    memcpy(helper, output_1, 32);
+    memcpy(helper, output_1, LT_HMAC_SHA256_HASH_LEN);  // Copy whole output of SHA256 HMAC.
     helper[32] = 2;
 
-    lt_hmac_sha256(tmp, 32, helper, 33, output_2);
+    lt_hmac_sha256(tmp, sizeof(tmp), helper, sizeof(helper), output_2);
 }